requirements.txt
tls-client
requests~=2.27.1
colorama~=0.4.4
svglib~=1.4.1
bs4~=0.0.1
beautifulsoup4~=4.10.0
reportlab~=3.6.12
discord.py
aiohttp